WebServer ausgebaut sodass er dem HTTP Protokoll entspricht

+++ MonsterTradingCards/Repositories/CardRepository.cs
@@ -36,8 +36,8 @@ namespace MonsterTradingCards.Repositories
                     command.AddParameterWithValue("name", DbType.String, card.Name);
 
                     command.AddParameterWithValue("damage", DbType.Double, card.Damage);
-                    command.AddParameterWithValue("element_type", DbType.Int32, (int)card.ElementType);
-                    command.AddParameterWithValue("card_type", DbType.Int32, (int)card.CardType);
+                    command.AddParameterWithValue("element_type", DbType.Int32, (int) card.ElementType);
+                    command.AddParameterWithValue("card_type", DbType.Int32, (int) card.CardType);
 
                     command.AddParameterWithValue("user_id", DbType.Guid, card.UserId);
                     command.AddParameterWithValue("is_in_deck", DbType.Boolean, card.IsInDeck);
@@ -199,8 +199,8 @@ namespace MonsterTradingCards.Repositories
                     command.AddParameterWithValue("name", DbType.String, card.Name);
 
                     command.AddParameterWithValue("damage", DbType.Double, card.Damage);
-                    command.AddParameterWithValue("element_type", DbType.Int32, card.ElementType);
-                    command.AddParameterWithValue("card_type", DbType.Int32, card.CardType);
+                    command.AddParameterWithValue("element_type", DbType.Int32, (int) card.ElementType);
+                    command.AddParameterWithValue("card_type", DbType.Int32, (int) card.CardType);
 
                     command.AddParameterWithValue("user_id", DbType.Guid, card.UserId);
                     command.AddParameterWithValue("is_in_deck", DbType.Boolean, card.IsInDeck);
